[iOS GPU] Implement view and reshape in metal shaders
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54107

The current implementation doesn't change the underlying texture's shape. This diff converts MPSImage from one shape to the other. The way we'll do it is we implement this as an elementwise kernel. We have a thread grid of size (N2, C2, H2, W2) with a thread for each output element, and we compute the "linear index" of the output element, and convert it to the equivalent "linear index" of the input element. This is the known as sub2ind/ind2sub conversion in MATLAB, ravel_multi_index in numpy, etc. https://github.com/cupy/cupy/blob/a08841a8e18d285652df0f7a2a25f5ae5bc1a897/cupy/indexing/generate.py#L301-L304 is a clean generic version of ind2sub.
ghstack-source-id: 124113407

Differential Revision: [D27074719](https://our.internmc.facebook.com/intern/diff/D27074719/)